Refactor mesh loading to support multiple file types.
